antigravity-whisper-stt/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
openai-whisper
aiofiles==23.2.1
pydantic==2.6.0
python-dotenv==1.0.1
moviepy==1.0.3
ffmpeg-python==0.2.0
